# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.CreateConversationRequest:
    properties:
      character_id:
        type: integer
    required:
    - character_id
    type: object
  models.SendMessageRequest:
    properties:
      audio_url:
        type: string
      content:
        type: string
    required:
    - content
    type: object
  models.UserLogin:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.UserRegister:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 基于AI的角色扮演聊天系统API文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: AI角色扮演聊天API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: 用户登录获取JWT token
      parameters:
      - description: 用户登录信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 认证失败
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 用户登录
      tags:
      - 认证
  /auth/logout:
    post:
      consumes:
      - application/json
      description: 用户登出，清除JWT token
      produces:
      - application/json
      responses:
        "200":
          description: 登出成功
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 用户登出
      tags:
      - 认证
  /auth/me:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的详细信息
      produces:
      - application/json
      responses:
        "200":
          description: 用户信息
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 用户不存在
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取当前用户信息
      tags:
      - 认证
  /auth/register:
    post:
      consumes:
      - application/json
      description: 创建新用户账户
      parameters:
      - description: 用户注册信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserRegister'
      produces:
      - application/json
      responses:
        "201":
          description: 注册成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 用户注册
      tags:
      - 认证
  /characters:
    get:
      consumes:
      - application/json
      description: 获取所有可用的AI角色列表
      produces:
      - application/json
      responses:
        "200":
          description: 角色列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取角色列表
      tags:
      - 角色
  /characters/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取特定角色的详细信息
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 角色详情
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 角色不存在
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取角色详情
      tags:
      - 角色
  /characters/search:
    get:
      consumes:
      - application/json
      description: 根据关键词搜索角色
      parameters:
      - description: 搜索关键词
        in: query
        name: query
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 搜索结果
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 搜索角色
      tags:
      - 角色
  /conversations:
    get:
      consumes:
      - application/json
      description: 获取当前用户的所有对话会话
      produces:
      - application/json
      responses:
        "200":
          description: 对话列表
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取对话列表
      tags:
      - 对话
    post:
      consumes:
      - application/json
      description: 创建一个新的对话会话
      parameters:
      - description: 创建对话请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateConversationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 对话创建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 创建新对话
      tags:
      - 对话
  /conversations/{id}/messages:
    post:
      consumes:
      - application/json
      description: 向对话发送消息并获取AI回复
      parameters:
      - description: 对话ID
        in: path
        name: id
        required: true
        type: integer
      - description: 消息内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SendMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 消息发送成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 发送消息
      tags:
      - 对话
securityDefinitions:
  BearerAuth:
    description: JWT认证，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"